a_DebugPanel/requirements.txt
kivy>=2.1.0
pyserial>=3.5
